requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
orjson==3.10.7
